feat: <Static> compatibility + refactor
