+++ packages/core/src/translate/downloadFileBatch.ts
@@ -7,7 +7,7 @@ import { TranslationRequestConfig } from '../types';
 import {
   DownloadFileBatchOptions,
   DownloadFileBatchResult,
-} from '../types-dir/downloadFileBatch';
+} from '../types-dir/api/downloadFileBatch';
 import generateRequestHeaders from './utils/generateRequestHeaders';
 import { decode } from '../utils/base64';
