+++ packages/core/src/types-dir/api/checkFileTranslations.ts
