+++ packages/core/src/types-dir/api/downloadFileBatch.ts
