+++ packages/core/src/types-dir/api/file.ts
