+++ packages/core/src/types-dir/api/translateMany.ts
