+++ packages/core/src/types-dir/api/uploadFiles.ts
@@ -1,4 +1,4 @@
-import { DataFormat } from './content';
+import { DataFormat } from '../jsx/content';
 import { FileFormat } from './file';
 
 export type FileUpload = {
