+++ packages/core/src/types-dir/transformations.ts
@@ -1,0 +1,26 @@
+/**
+ * Transformations are made from a prefix and a suffix.
+ */
+export type Transformation =
+  | 'translate-client'
+  | 'translate-server'
+  | 'translate-runtime'
+  | 'variable-variable'
+  | 'variable-currency'
+  | 'variable-datetime'
+  | 'variable-number'
+  | 'variable-static'
+  | 'plural'
+  | 'branch';
+export type TransformationPrefix =
+  | 'translate'
+  | 'variable'
+  | 'plural'
+  | 'branch'
+  | 'fragment';
+export type VariableTransformationSuffix =
+  | 'variable'
+  | 'number'
+  | 'datetime'
+  | 'currency'
+  | 'static';
